requirements.txt
cycler==0.10.0
h5py==2.7.1
matplotlib==2.0.0
olefile==0.44
pillow==5.1.0
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.3
pyyaml==3.12
scipy==0.19.0
setuptools==27.2.0
six==1.11.0
wheel==0.29.0
keras==2.1.5
protobuf==3.5.2.post1
pydot-ng==1.0.0
tensorflow==1.4.0
numpy==1.14.2
fire==0.1.3
